requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
async-timeout==5.0.1
asyncio==3.4.3
attrs==24.3.0
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
propcache==0.2.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
six==1.17.0
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.15.2

flask~=3.0.3
setuptools~=68.2.0
requests~=2.32.3